Moving to "controling variables" namespace
