+++ AnthillNet/NetworkHost.cs
@@ -1,4 +1,5 @@
 ﻿using AnthillNet.Core;
+using AnthillNet.CVar;
 using System;
 using System.Collections.Generic;
 using System.Net;
